requirements.txt
pandas
scipy
streamlit
plotly-express